+++ recipes/db_restore_mongo_backups_setup.rb
@@ -8,14 +8,20 @@ cookbook_file "/root/backup_management.rb" do
   mode   '0755'
 end
 
-cron_commands = [
-  "ruby /root/backup_management.rb /mnt/postgresbackups/backups " +
-  "#{ node['environment_name'] } #{ get_current_applications({'database' => 'mongodb'}).join(',') } " +
-  "#{ Chef::EncryptedDataBagItem.load( node.chef_environment, 'chef_passwords', node['secret']).to_hash["mongo_pass"] } mongodb " +
-  "> /root/restore_mongodb.log 2>&1"
-]
+
+cron_commands = []
 
 if data_bag_item( node.chef_environment, 'config').to_hash[node['environment_name']]['restore_backups']
+  node['loaded_applications'].each_key do |app_role_name|
+    next unless has_repo_hash?(app_role_name)
+    next unless repo_hash(app_role_name)['database'] == 'mongodb'
+
+    cron_commands << "ruby /root/backup_management.rb /mnt/postgresbackups/backups " +
+      "#{ node['environment_name'] } #{ get_current_applications({'database' => 'mongodb'}).join(',') } " +
+      "#{ Chef::EncryptedDataBagItem.load( node.chef_environment, 'chef_passwords', node['secret']).to_hash["mongo_pass"] } " +
+      "#{ repo_hash(app_role_name)['application_database_user'] } mongodb > /root/restore_mongodb.log 2>&1"
+  end
+
   cron "restore_from_production_mongo_database" do
     minute  "45"
     hour    "11"